fix(comment)!: fall back to using trimmed comment markers
Currently comment detection, addition, and removal are done by matching
'commentstring' exactly (possibly forced to single padding). This has
the downside when users want to add comment markers with space (like
with `-- %s` commentstring) but also be able to uncomment lines that do
not contain space (like `--aaa`).

Resolve this with the following approach:
  - Line is commented if it matches 'commentstring' with trimmed parts.
  - Adding comment is 100% relying on 'commentstring' parts (as is now).
  - Removing comment is first trying exact 'commentstring' parts with
    fallback on trying its trimmed parts.

This aligns behavior with Neovim's built-in commenting.
See neovim/neovim#28938.